fastapi_app/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
paho-mqtt==2.1.0
SQLAlchemy==2.0.34
sse-starlette==2.1.2
pydantic==2.9.2
requests==2.32.3
python-multipart==0.0.9
ultralytics==8.2.103
opencv-python-headless==4.10.0.84
numpy>=1.24
python-multipart==0.0.9
